Tuotteen nimi ja valmistaja tietokantaan
Plus muita satunnaisia muutoksia.

- Ajax_reqs : funktio, joka tallentaa tuotteen nimen ja valmistajan.
Testattu toimivan. Plus whitespace muutoksia.
- User-luokka : korjaus. Poistettu taulun kolumni oli jäänyt koodiin.
-Noutolista_pdf_luonti.php : ei valmis, ajattelin vain lisätä sen
GitHubiin.
- Omat_tiedot : whitespace, ja hieman koodin yhtenäistämistä.
- Tietokanta : Uusi taulu laskunumerolle (ei vielä käytössä). Pari uutta
riviä ostotilauskirjaan kun olin siinä. Constraint nimimuutoksia, ja
whitespacemuutoksia.
Nuo constraintit on nimetty "fk_fkTaulu_originalTaulu" tyylin mukaan.
- Tuotehaku : ostoskorin lisäyksessä tuotteen nimi ja valmistaja. Uusi
lähde jQuerylle. Whitespace.
- Tuotemodal : oletan, että kukaan ei hirveän usein käy tuolla, joten
otin toisen suppressin pois. Y'know, jos joku siellä joskus käy.
- Yp_muokkaa_asiakas : Ups, mitenkäs tuo tuonne unohtu? Sori, debugging
unohtu päälle.
- Yp_tuotteet : tuotteen lisäyksessä nimen ja valmistajan tallennus
kantaan. Muokatessa tuotetta ei muistaakseen testauksessa toiminu, mutta
saatan vielä hieman enemmän siivota koodia täällä. Whitespace.
Tyylimuutoksia (uusi header-row)

+++ luokat/laskutiedot.class.php
@@ -87,7 +87,7 @@ class Laskutiedot {
 			$tuote->a_hinta = round($row->maksettu_hinta, 2);
 			$tuote->a_hinta_ilman_alv = $row->pysyva_hinta * (1 - $row->pysyva_alennus);
 			$tuote->alv_prosentti = (int)((float)$row->pysyva_alv * 100); // 0.24 => 24
-			$tuote->alennus = (float)$row->pysyva_alennus * 100; // 0.24 => 24
+			$tuote->alennus = (int)((float)$row->pysyva_alennus * 100); // 0.24 => 24
 			$tuote->kpl_maara = $row->kpl;
 			$tuote->summa = ($row->maksettu_hinta * $row->kpl);
 			$this->tuotteet[] = $tuote;
